refactor: 今までホームページだったところを/searchに移した
